Extracting registration logic from the controller
